// src/api/medicine/models/MedicineOrderEntry.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Medicine Inventory API
 * Medicine inventory management for Web-In-Cloud system
 *
 * The version of the OpenAPI document: 1.0.0
 * Contact: your_email@stuba.sk
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { exists, mapValues } from '../runtime';
import type { Status } from './Status';
import {
    StatusFromJSON,
    StatusFromJSONTyped,
    StatusToJSON,
} from './Status';

/**
 * 
 * @export
 * @interface MedicineOrderEntry
 */
export interface MedicineOrderEntry {
    /**
     * Unique id of the entry in this medicine inventory
     * @type {string}
     * @memberof MedicineOrderEntry
     */
    id: string;
    /**
     * Name of medicine in medicine inventory
     * @type {string}
     * @memberof MedicineOrderEntry
     */
    name?: string;
    /**
     * Unique identifier of the medicine known to Web-In-Cloud system
     * @type {string}
     * @memberof MedicineOrderEntry
     */
    medicineId: string;
    /**
     * Medicine count in the ambulance medicine inventory. It is a number of packages in the medicine inventory for the given ambulance.
     * @type {number}
     * @memberof MedicineOrderEntry
     */
    count: number;
    /**
     * 
     * @type {Status}
     * @memberof MedicineOrderEntry
     */
    status: Status;
}

/**
 * Check if a given object implements the MedicineOrderEntry interface.
 */
export function instanceOfMedicineOrderEntry(value: object): boolean {
    let isInstance = true;
    isInstance = isInstance && "id" in value;
    isInstance = isInstance && "medicineId" in value;
    isInstance = isInstance && "count" in value;
    isInstance = isInstance && "status" in value;

    return isInstance;
}

export function MedicineOrderEntryFromJSON(json: any): MedicineOrderEntry {
    return MedicineOrderEntryFromJSONTyped(json, false);
}

export function MedicineOrderEntryFromJSONTyped(json: any, ignoreDiscriminator: boolean): MedicineOrderEntry {
    if ((json === undefined) || (json === null)) {
        return json;
    }
    return {
        
        'id': json['id'],
        'name': !exists(json, 'name') ? undefined : json['name'],
        'medicineId': json['medicineId'],
        'count': json['count'],
        'status': StatusFromJSON(json['status']),
    };
}

export function MedicineOrderEntryToJSON(value?: MedicineOrderEntry | null): any {
    if (value === undefined) {
        return undefined;
    }
    if (value === null) {
        return null;
    }
    return {
        
        'id': value.id,
        'name': value.name,
        'medicineId': value.medicineId,
        'count': value.count,
        'status': StatusToJSON(value.status),
    };
}